+++ log/output-react/src/lib/global.d.ts
@@ -11,7 +11,6 @@ declare global {
     setAllEntriesWhenPossible: (
       allEntries: Entry[],
       newExpanded: string[],
-      consoleEntries: EntryConsole[],
       updatedFromIndex: number,
     ) => void;
   }