+++ log/src/robocorp/log/_obj_info_repr.py
@@ -22,7 +22,7 @@ def get_obj_type_and_repr(obj: Any) -> Tuple[str, str]:
         except (Exception, RecursionError) as e:
             r = f"<error getting repr: {e}>"
 
-        max_size = _config._general_log_config.get_max_value_repr_size()
+        max_size = _config._general_log_config.max_value_repr_size
         if len(r) > max_size:
             diff = len(r) - max_size
             r = f"{r[:max_size]} <clipped {diff} chars>"