+++ tasks/src/robocorp/tasks/_protocols.py
@@ -29,13 +29,17 @@ def check_implements(x: T) -> T:
     return x
 
 
+# Note: this is a bit messy as we're mixing task states with log levels.
+# Note2: This is for the log.html and not really for user APIs.
 class Status:
     NOT_RUN = "NOT_RUN"  # Initial status for a task which is not run.
-    PASS = "PASS"
-    ERROR = "ERROR"
-    FAIL = "FAIL"
-    INFO = "INFO"
-    WARN = "WARN"
+    PASS = "PASS"  # Used for task pass
+    FAIL = "FAIL"  # Used for task failure
+
+    ERROR = "ERROR"  # log.critical
+    INFO = "INFO"  # log.info
+    WARN = "WARN"  # log.warn
+    DEBUG = "DEBUG"  # log.debug
 
 
 @dataclass